style for Users component refactored
