+++ src/users/users-style.ts
@@ -1,0 +1,8 @@
+import { StyleSheet } from 'react-native';
+
+export const styles = StyleSheet.create({
+  item: {
+    borderBottomColor: 'black',
+    borderBottomWidth: 0.5,
+  },
+});
